+++ CHANGELOG.md
@@ -20,6 +20,8 @@ See this project's [releases](/../../../releases).
 
 ## [Unreleased]
 
+## [1.1.5] - 2024-02-21
+
 ### Added
 
 - Reference codeblocks for referencing local and external files
@@ -310,7 +312,8 @@ See this project's [releases](/../../../releases).
 - Updated [README](README.md).
 - Removed `none` language from supported languages since it is used by obsidian for codeblocks without a language
 
-[Unreleased]: /../../compare/1.1.4...HEAD
+[Unreleased]: /../../compare/1.1.5...HEAD
+[1.1.5]: /../../compare/1.1.4...1.1.5
 [1.1.4]: /../../compare/1.1.3...1.1.4
 [1.1.3]: /../../compare/1.1.2...1.1.3
 [1.1.2]: /../../compare/1.1.1...1.1.2